backend/user_management/utils/auth_helpers.py
import jwt
import os
from datetime import datetime, timedelta
from functools import wraps
from bson import ObjectId
from flask import request, jsonify

JWT_SECRET = os.getenv("JWT_SECRET", "jwt-secret-key")
JWT_EXPIRY_HOURS = int(os.getenv("JWT_EXPIRY_HOURS", 24))


def generate_token(user_id: str, role: str) -> str:
    payload = {
        "sub": str(user_id),
        "role": role,
        "exp": datetime.utcnow() + timedelta(hours=JWT_EXPIRY_HOURS),
    }
    return jwt.encode(payload, JWT_SECRET, algorithm="HS256")


def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        from ..database import get_db
        from ..models.user import User

        auth = request.headers.get("Authorization", "")
        if not auth.startswith("Bearer "):
            return jsonify({"error": "Token missing"}), 401
        token = auth.split(" ", 1)[1]
        try:
            payload = jwt.decode(token, JWT_SECRET, algorithms=["HS256"])
        except jwt.ExpiredSignatureError:
            return jsonify({"error": "Token expired"}), 401
        except jwt.InvalidTokenError:
            return jsonify({"error": "Invalid token"}), 401

        db = get_db()
        try:
            doc = db.users.find_one({"_id": ObjectId(payload["sub"])})
        except Exception:
            return jsonify({"error": "Invalid token payload"}), 401

        if not doc:
            return jsonify({"error": "User not found"}), 401
        return f(User(doc), *args, **kwargs)
    return decorated


def admin_required(f):
    @wraps(f)
    def decorated(current_user, *args, **kwargs):
        if current_user.role != "admin":
            return jsonify({"error": "Admin access required"}), 403
        return f(current_user, *args, **kwargs)
    return decorated
